Remove username update, add auto username updating and username resolution, add account links, add max entries limit

+++ follow_manager.sh
@@ -5,6 +5,31 @@
 
 CONFIG_DIR="$HOME/.github-follow-manager"
 ENV_FILE="$CONFIG_DIR/.env"
+GITHUB_API_URL="https://api.github.com"
+GITHUB_URL="https://github.com"
+MAX_ENTRIES=100
+
+get_current_username() {
+    curl -s -u "anonymous:$GITHUB_TOKEN" "$GITHUB_API_URL/user" | jq -r '.login'
+}
+
+check_current_username() {
+    username=$(get_current_username)
+    if [ "$username" != "$GITHUB_USER" ]; then
+        token=$(grep GITHUB_TOKEN "$ENV_FILE" | cut -d'=' -f2)
+        update_env_file $username $token
+    fi
+}
+
+update_env_file() {
+    if [ -z "$1" ] || [ -z "$2" ]; then
+        echo "GitHub username and token are required to update .env file."
+        exit 1
+    fi
+    echo "Updating .env file..."
+    echo "GITHUB_USER=$1" > "$ENV_FILE"
+    echo "GITHUB_TOKEN=$2" >> "$ENV_FILE"
+}
 
 if ! command -v jq &> /dev/null; then
     echo "jq is required to run this script. Please install jq first."
@@ -21,22 +46,35 @@ fi
 if [ ! -f "$ENV_FILE" ]; then
     echo ".env file not found! Creating one..."
     mkdir -p "$CONFIG_DIR"
-    read -p "Enter your GitHub username: " GITHUB_USER
     read -sp "Enter your GitHub token: " GITHUB_TOKEN
     echo
-    echo "GITHUB_USER=$GITHUB_USER" > "$ENV_FILE"
-    echo "GITHUB_TOKEN=$GITHUB_TOKEN" >> "$ENV_FILE"
+    GITHUB_USER=$(get_current_username)
+    if [ -z "$GITHUB_USER" ]; then
+        echo "Failed to fetch GitHub username. Please try again later."
+        exit 1
+    fi
+    update_env_file $GITHUB_USER $GITHUB_TOKEN
     echo ".env file created with your GitHub credentials."
 else
     export $(cat "$ENV_FILE" | xargs)
 fi
 
-if [ -z "$GITHUB_USER" ] || [ -z "$GITHUB_TOKEN" ]; then
-    echo "GitHub username or token is not set correctly. Please check your .env file in $CONFIG_DIR."
+if [ -z "$GITHUB_TOKEN" ]; then
+    echo "GitHub token is not set correctly. Please check your .env file in $CONFIG_DIR."
     exit 1
 fi
 
-GITHUB_API_URL="https://api.github.com"
+if [ -z "$GITHUB_USER" ]; then
+    echo "GitHub username not found in .env file. Trying to fetch it..."
+    GITHUB_USER=$(get_current_username)
+    if [ -z "$GITHUB_USER" ]; then
+        echo "Failed to fetch GitHub username. Please check your .env file in $CONFIG_DIR."
+        exit 1
+    fi
+    update_env_file $GITHUB_USER $GITHUB_TOKEN
+else
+    check_current_username
+fi
 
 follow_back() {
     followers=$(curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" "$GITHUB_API_URL/users/$GITHUB_USER/followers" | jq -r '.[].login')
@@ -54,16 +92,29 @@ follow_back() {
     echo "You have $(echo "$followers" | wc -l) followers and you are not following back $(echo "${not_following_back[@]}" | wc -w) of them."
     if [ ${#not_following_back[@]} -eq 0 ]; then
         echo "You are already following back all followers."
-        exit 0
+        return
     fi
 
-    echo "Do you want to follow back all followers? (y/n)"
+    if [ ${#not_following_back[@]} -gt $MAX_ENTRIES ]; then
+        echo "You have too many followers to follow back to show them all."
+    else
+        echo "Users not followed back:"
+        for user in "${not_following_back[@]}"; do
+            echo -e "\e]8;;$GITHUB_URL/$user\a$user\e]8;;\a"
+        done
+    fi
+
+    echo "Do you want to follow back all followers? [y/n]"
     read -r confirm
     if [ "$confirm" != "y" ]; then
+        if [ ${#not_following_back[@]} -gt $MAX_ENTRIES ]; then
+            echo "You have too many followers to follow back manually."
+            return
+        fi
         echo "Select a user to follow back or choose 0 to abort:"
         select user in "${not_following_back[@]}"; do
             if [ -n "$user" ]; then
-                echo "Following $user..."
+                echo -e "Following \e]8;;$GITHUB_URL/$user\a$user\e]8;;\a..."
                 curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" -X PUT "$GITHUB_API_URL/user/following/$user" > /dev/null
                 break
             elif [ "$REPLY" -eq 0 ]; then
@@ -76,10 +127,10 @@ follow_back() {
         done
         exit 1
     fi
+
     for user in $followers; do
         echo "Following $user..."
         curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" -X PUT "$GITHUB_API_URL/user/following/$user" > /dev/null
-
     done
     echo "All followers have been followed back."
 }
@@ -95,24 +146,31 @@ unfollow_non_followers() {
         fi
     done
     echo "There are $(echo "$following" | wc -l) users you are following and $(echo "${non_followers[@]}" | wc -w) of them are not following you back."
-    echo "Users who are not following you back:"
-    
     if [ ${#non_followers[@]} -eq 0 ]; then
         echo "No users to unfollow."
-        exit 0
+        return
     fi
 
-    for user in "${non_followers[@]}"; do
-        echo "$user"
-    done
+    if [ ${#non_followers[@]} -gt $MAX_ENTRIES ]; then
+        echo "You have too many users to unfollow to show them all."
+    else
+        echo "Users who are not following you back:"
+        for user in "${non_followers[@]}"; do
+            echo -e "\e]8;;$GITHUB_URL/$user\a$user\e]8;;\a"
+        done
+    fi
     
-    echo "Do you want to unfollow all users who are not following you back? (y/n)"
+    echo "Do you want to unfollow all users who are not following you back? [y/n]"
     read -r confirm
     if [ "$confirm" != "y" ]; then
+        if [ ${#non_followers[@]} -gt $MAX_ENTRIES ]; then
+            echo "You have too many users to unfollow manually."
+            return
+        fi
         echo "Select a user to unfollow or choose 0 to abort:"
         select user in "${non_followers[@]}"; do
             if [ -n "$user" ]; then
-                echo "Unfollowing $user..."
+                echo "Unfollowing \e]8;;$GITHUB_URL/$user\a$user\e]8;;\a..."
                 curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" -X DELETE "$GITHUB_API_URL/user/following/$user" > /dev/null
                 break
             elif [ "$REPLY" -eq 0 ]; then
@@ -136,13 +194,11 @@ unfollow_non_followers() {
 list_followers() {
     echo "Followers:"
     curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" "$GITHUB_API_URL/users/$GITHUB_USER/followers" | jq -r '.[].login'
-    exit 0
 }
 
 list_following() {
     echo "Following:"
     curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" "$GITHUB_API_URL/users/$GITHUB_USER/following" | jq -r '.[].login'
-    exit 0
 }
 
 list_not_following_back() {
@@ -151,10 +207,9 @@ list_not_following_back() {
     followers=$(curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" "$GITHUB_API_URL/users/$GITHUB_USER/followers" | jq -r '.[].login')
     for user in $following; do
         if ! echo "$followers" | grep -q "$user"; then
-            echo "$user"
+            echo -e "\e]8;;$GITHUB_URL/$user\a$user\e]8;;\a"
         fi
     done
-    exit 0
 }
 
 list_not_followed_back() {
@@ -162,15 +217,9 @@ list_not_followed_back() {
     followers=$(curl -s -u "$GITHUB_USER:$GITHUB_TOKEN" "$GITHUB_API_URL/users/$GITHUB_USER/followers" | jq -r '.[].login')
     for user in $followers; do
         if ! echo "$following" | grep -q "$user"; then
-            echo "$user"
+            echo -e "\e]8;;$GITHUB_URL/$user\a$user\e]8;;\a"
         fi
     done
-}
-
-update_username() {
-    read -p "Enter your GitHub username: " GITHUB_USER
-    echo "GITHUB_USER=$GITHUB_USER" > "$ENV_FILE"
-    echo "GitHub username updated successfully."
 }
 
 usage() {
@@ -182,7 +231,6 @@ usage() {
     echo "  list-following: List all following users"
     echo "  list-not-following-back: List users who are not following you back"
     echo "  list-not-followed-back: List users who you are not following back"
-    echo "  update-username: Update your GitHub username"
 }
 
 display_menu() {
@@ -194,7 +242,6 @@ display_menu() {
     echo "4) List following"
     echo "5) List not-following back"
     echo "6) List not-followed back"
-    echo "7) Update username"
 }
 
 if [ $# -eq 0 ]; then
@@ -224,9 +271,6 @@ if [ $# -eq 0 ]; then
             6)
                 list_not_followed_back
                 ;;
-            7)
-                update_username
-                ;;
             *)
                 echo "Invalid option. Please try again."
                 ;;
@@ -252,9 +296,6 @@ case "$1" in
         ;;
     list-not-followed-back)
         list_not_followed_back
-        ;;
-    update-username)
-        update_username
         ;;
     *)
         usage
